backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
llama-index>=0.12.0
llama-index-llms-openai-like>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-chroma>=0.4.0
chromadb>=0.5.0
pymupdf>=1.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlmodel>=0.0.22
pytesseract>=0.3.10
Pillow>=10.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.1.0,<4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
